requirements.txt
langgraph
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai
streamlit
chromadb
sqlalchemy
psycopg2-binary
pymysql
db-sqlite3
openai
pandas
numpy
plotly
matplotlib
seaborn
python-dotenv
tqdm
ipykernel
jupyter